Fix issue #1476
