+++ src/hip/general_tensor_reorder_sol.cpp
@@ -52,6 +52,17 @@ static inline std::string GetKernelNameType(std::size_t type_size)
     MIOPEN_THROW("data type not supported");
 }
 
+static inline std::string GetKernelFileName(std::size_t data_size,
+                                            const GeneralReorderParam* kparam)
+{
+    if(kparam == nullptr)
+        MIOPEN_THROW("Memory access fault, kparam is a nullptr");
+    std::ostringstream kernel_file_name;
+    kernel_file_name << "general_tensor_reorder_" << kparam->tile_x << "x" << kparam->tile_y << "_";
+    kernel_file_name << GetKernelNameType(data_size) << ".cpp";
+    return kernel_file_name.str();
+}
+
 static inline std::string GetKernelName(std::size_t data_size,
                                         uint32_t order_0,
                                         uint32_t order_1,
@@ -140,9 +151,10 @@ solver::KernelInfo GenericReorderSolutionImpl::GetKernelInfo() const
                          (block_size * kernel_param_heuristic.tile_x);
     std::size_t grid_size = dim_total;
 
-    std::string kernel_name = GetKernelName();
+    std::string kernel_name      = GetKernelName();
+    std::string kernel_file_name = GetKernelFileName();
     solver::KernelInfo kernel;
-    kernel.kernel_file = "general_tensor_reorder.cpp";
+    kernel.kernel_file = kernel_file_name;
     kernel.kernel_name = kernel_name;
     kernel.g_wk.clear();
     kernel.g_wk.push_back(grid_size * block_size);
@@ -189,6 +201,12 @@ std::vector<OpKernelArg> GenericReorderSolutionImpl::GetKernelArg() const
     opArgs.emplace_back(static_cast<uint32_t>(magic_stride2.shift));
 
     return opArgs;
+}
+
+std::string GenericReorderSolutionImpl::GetKernelFileName() const
+{
+    return tensor_reorder::GetKernelFileName(miopen::GetTypeSize(data_type),
+                                             &kernel_param_heuristic);
 }
 
 std::string GenericReorderSolutionImpl::GetKernelName() const
