--- src/kernels/gpu_general_tensor_reorder_kernel/order.hpp
@@ -1,43 +1,0 @@
-/*******************************************************************************
- *
- * MIT License
- *
- * Copyright (c) 2020-2022 Advanced Micro Devices, Inc.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- *
- *******************************************************************************/
-#include <cstdint>
-#ifndef ORDER_HPP
-#define ORDER_HPP
-
-template <int... Is>
-struct order
-{
-    static constexpr std::size_t m_size = sizeof...(Is);
-    // the last dummy element is to prevent compiler complain about empty array, when mSize = 0
-    static constexpr int m_data[m_size + 1] = {Is..., 0};
-
-    __host__ __device__ static constexpr uint64_t size() { return m_size; }
-
-    __host__ __device__ static constexpr uint64_t get_size() { return size(); }
-
-    __host__ __device__ static constexpr int at(int I) { return m_data[I]; }
-};
-#endif
